requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
Pygments==2.18.0
PySDL2==0.9.16
pysdl2-dll==2.30.2
pytz==2024.2
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.3.7
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
